Converted perft results to long. Fixed typo. All tests are successful!

+++ src/main/java/de/janfrase/blunder/Main.java
@@ -32,7 +32,7 @@ public class Main {
 
         // int nodes = PerftRunner.perft(3);
 
-        int nodes =
+        long nodes =
                 PerftRunner.perft(
                         4, "r3k2r/p1ppqpb1/bn2pnp1/3PN3/1p2P3/2N2Q1p/PPPBBPPP/R3K2R w KQkq - 0 1");
